RichTextEditor: Hashtag visual transformation performance improvement
Changed the algorithm for tagging hashtags in the text that would do a single pass. The implementation is still a tad slow since it has to re-compose everytime the text changes. Perhaps we could look into a storing mechanism of all the hash tags and only trigger this when the list of hash tags look different